requirements.txt
git+https://github.com/huggingface/transformers
pandas==2.2.3
openpyxl==3.1.5
google-generativeai==0.8.2
python-dotenv==1.0.1

